Changed defname internal (numerical) value system. Changes to CResourceID, to have a separate word variable to hold the "page" number, instead of having it inside the uid. Re-added CResourceIDBase, which DOESN'T have the page info.
WARNING: the following change could break some hackish expedient used in scripts! Read carefully!
- Changed: Defname numeric value has a different byte order now. For example, <hval s_clumsy> returned 0ce000001, so you could retrieve the spell number by doing <hval s_clumsy&(~0ce000000)>,	which removed the resource type bits. Now <hval s_clumsy> returns 082700001, so you can't use the previous code anymore. Use RESOURCEINDEX instead. This change was needed because too much informations were stored inside the UID, and its last bytes (used to store the resource type) overlapped with some internal flags.
- Added: <RESOURCEINDEX defname> to retrieve the resource index from the defname, ignoring the resource type (Chardef, Spell, Itemdef, Template...) bits. As example, <RESOURCEINDEX s_clumsy> returns 1, which is the spell number (id).
- Added: <RESOURCETYPE defname>, which gets the resource type (RES_ITEMDEF, RES_SPAWN, etc) from the defname. The list of the resource type numbers is in the X scriptpack in core/defs.scp.
- Changed: REGIONTYPE block could not accept landtile types with index > 255. Extended this limit to 65534. This also extends the maximum number of pages in a book by the same amount.

+++ src/game/clients/CClientEvent.cpp
@@ -151,13 +151,13 @@ void CClient::Event_Item_Dye( CUID uid, HUE_TYPE wHue ) // Rehue an item
 }
 
 
-void CClient::Event_Tips( word i) // Tip of the day window
+void CClient::Event_Tips(word i) // Tip of the day window
 {
 	ADDTOCALLSTACK("CClient::Event_Tips");
-	if (i==0)
-		i=1;
+	if (i == 0)
+		i = 1;
 	CResourceLock s;
-	if ( g_Cfg.ResourceLock( s, CResourceID( RES_TIP, i )) == false )
+	if ( g_Cfg.ResourceLock( s, CResourceID( RES_TIP, (int)i )) == false )
 	{
 		// requested tip was not found, default to tip 1 if possible
 		if ( i == 1 || ( g_Cfg.ResourceLock( s, CResourceID( RES_TIP, 1 )) == false ))
